@@ -17,12 +17,36 @@ from azure.ai.ml.entities._job._input_output_helpers import from_rest_data_outpu
 from azure.ai.ml.entities._job.automl.tabular.automl_tabular import AutoMLTabular
 from azure.ai.ml.entities._job.automl.tabular.featurization_settings import TabularFeaturizationSettings
 from azure.ai.ml.entities._job.automl.tabular.limit_settings import TabularLimitSettings
-from azure.ai.ml.entities._job.automl.training_settings import ClassificationTrainingSettings
+from azure.ai.ml.entities._job.automl.training_settings import (  # noqa: F401 # pylint: disable=unused-import
+    ClassificationTrainingSettings,
+    TrainingSettings,
+)
 from azure.ai.ml.entities._util import load_from_dict
 
 
 class ClassificationJob(AutoMLTabular):
-    """Configuration for AutoML Classification Job."""
+    """Configuration for AutoML Classification Job.
+
+    :keyword primary_metric: The primary metric to use for optimization, defaults to None
+    :paramtype primary_metric: typing.Optional[str]
+    :keyword positive_label: Positive label for binary metrics calculation, defaults to None
+    :paramtype positive_label: typing.Optional[str]
+    :keyword kwargs: Additional keyword arguments
+    :paramtype kwargs: typing.Any
+    :keyword featurization: Featurization settings. Defaults to None.
+    :paramtype featurization: typing.Optional[TabularFeaturizationSettings]
+    :keyword limits: Limits settings. Defaults to None.
+    :paramtype limits: typing.Optional[TabularLimitSettings]
+    :keyword training: Training settings. Defaults to None.
+    :paramtype training: typing.Optional[TrainingSettings]
+    :return: An instance of ClassificationJob object.
+    :rtype: ~azure.ai.ml.entities.automl.ClassificationJob
+    :raises ValueError: If primary_metric is not a valid primary metric
+    :raises ValueError: If positive_label is not a valid positive label
+    :raises ValueError: If featurization is not a valid featurization settings
+    :raises ValueError: If limits is not a valid limits settings
+    :raises ValueError: If training is not a valid training settings
+    """
 
     _DEFAULT_PRIMARY_METRIC = ClassificationPrimaryMetrics.ACCURACY
 
@@ -35,12 +59,25 @@ class ClassificationJob(AutoMLTabular):
     ) -> None:
         """Initialize a new AutoML Classification task.
 
-        :param primary_metric: The primary metric to use for optimization
-        :type primary_metric: str, optional
-        :param positive_label: Positive label for binary metrics calculation.
-        :type positive_label: str, optional
-        :param kwargs: Job-specific arguments
-        :type kwargs: dict
+        :keyword primary_metric: The primary metric to use for optimization, defaults to None
+        :paramtype primary_metric: typing.Optional[str]
+        :keyword positive_label: Positive label for binary metrics calculation, defaults to None
+        :paramtype positive_label: typing.Optional[str]
+        :keyword kwargs: Additional keyword arguments
+        :paramtype kwargs: typing.Any
+        :keyword featurization: featurization settings. Defaults to None.
+        :paramtype featurization: typing.Optional[TabularFeaturizationSettings]
+        :keyword limits: limits settings. Defaults to None.
+        :paramtype limits: typing.Optional[TabularLimitSettings]
+        :keyword training: training settings. Defaults to None.
+        :paramtype training: typing.Optional[TrainingSettings]
+        :return: An instance of ClassificationJob object.
+        :rtype: ~azure.ai.ml.entities.automl.ClassificationJob
+        :raises ValueError: If primary_metric is not a valid primary metric
+        :raises ValueError: If positive_label is not a valid positive label
+        :raises ValueError: If featurization is not a valid featurization settings
+        :raises ValueError: If limits is not a valid limits settings
+        :raises ValueError: If training is not a valid training settings
         """
         # Extract any task specific settings
         featurization = kwargs.pop("featurization", None)
@@ -60,10 +97,20 @@ class ClassificationJob(AutoMLTabular):
 
     @property
     def primary_metric(self):
+        """The primary metric to use for optimization.
+
+        :return: The primary metric to use for optimization.
+        :rtype: typing.Union[str, ClassificationPrimaryMetrics]
+        """
         return self._primary_metric
 
     @primary_metric.setter
     def primary_metric(self, value: Union[str, ClassificationPrimaryMetrics]):
+        """The primary metric to use for optimization setter.
+
+        :param value: Primary metric to use for optimization.
+        :type value: typing.Union[str, ClassificationPrimaryMetrics]
+        """
         # TODO: better way to do this
         if is_data_binding_expression(str(value), ["parent"]):
             self._primary_metric = value
@@ -76,9 +123,19 @@ class ClassificationJob(AutoMLTabular):
 
     @property
     def training(self) -> ClassificationTrainingSettings:
+        """Training Settings for AutoML Classification Job.
+
+        :return: Training settings used for AutoML Classification Job.
+        :rtype: ClassificationTrainingSettings
+        """
         return self._training or ClassificationTrainingSettings()
 
     def _to_rest_object(self) -> JobBase:
+        """Convert ClassificationJob object to a REST object.
+
+        :return: REST object representation of this object.
+        :rtype: JobBase
+        """
         classification_task = RestClassification(
             target_column_name=self.target_column_name,
             training_data=self.training_data,
@@ -122,6 +179,14 @@ class ClassificationJob(AutoMLTabular):
 
     @classmethod
     def _from_rest_object(cls, obj: JobBase) -> "ClassificationJob":
+        """Convert a REST object to ClassificationJob object.
+
+        :param obj: ClassificationJob in Rest format.
+        :type obj: JobBase
+        :return: ClassificationJob objects.
+        :rtype: ClassificationJob
+        """
+
         properties: RestAutoMLJob = obj.properties
         task_details: RestClassification = properties.task_details
 
@@ -139,9 +204,9 @@ class ClassificationJob(AutoMLTabular):
             "compute": properties.compute_id,
             "outputs": from_rest_data_outputs(properties.outputs),
             "resources": properties.resources,
-            "identity": _BaseJobIdentityConfiguration._from_rest_object(properties.identity)
-            if properties.identity
-            else None,
+            "identity": (
+                _BaseJobIdentityConfiguration._from_rest_object(properties.identity) if properties.identity else None
+            ),
             "queue_settings": properties.queue_settings,
         }
 
@@ -183,6 +248,17 @@ class ClassificationJob(AutoMLTabular):
         additional_message: str,
         **kwargs,
     ) -> "ClassificationJob":
+        """Load from a dictionary.
+
+        :param data: dictionary representation of the object.
+        :type data: typing.Dict
+        :param context: dictionary containing the context.
+        :type context: typing.Dict
+        :param additional_message: additional message to be added to the error message.
+        :type additional_message: str
+        :return: ClassificationJob object.
+        :rtype: ClassificationJob
+        """
         from azure.ai.ml._schema.automl.table_vertical.classification import AutoMLClassificationSchema
         from azure.ai.ml._schema.pipeline.automl_node import AutoMLClassificationNodeSchema
 
@@ -201,6 +277,13 @@ class ClassificationJob(AutoMLTabular):
 
     @classmethod
     def _create_instance_from_schema_dict(cls, loaded_data: Dict) -> "ClassificationJob":
+        """Create an instance from a schema dictionary.
+
+        :param loaded_data: dictionary containing the data.
+        :type loaded_data: typing.Dict
+        :return: ClassificationJob object.
+        :rtype: ClassificationJob
+        """
         loaded_data.pop(AutoMLConstants.TASK_TYPE_YAML, None)
         data_settings = {
             "training_data": loaded_data.pop("training_data"),
@@ -218,6 +301,13 @@ class ClassificationJob(AutoMLTabular):
         return job
 
     def _to_dict(self, inside_pipeline=False) -> Dict:  # pylint: disable=arguments-differ
+        """Convert the object to a dictionary.
+
+        :param inside_pipeline: whether the job is inside a pipeline or not, defaults to False
+        :type inside_pipeline: bool
+        :return: dictionary representation of the object.
+        :rtype: typing.Dict
+        """
         from azure.ai.ml._schema.automl.table_vertical.classification import AutoMLClassificationSchema
         from azure.ai.ml._schema.pipeline.automl_node import AutoMLClassificationNodeSchema
 
@@ -229,13 +319,30 @@ class ClassificationJob(AutoMLTabular):
         return schema_dict
 
     def __eq__(self, other):
+        """Returns True if both instances have the same values.
+
+        This method check instances equality and returns True if both of
+            the instances have the same attributes with the same values.
+
+        :param other: An instance of ClassificationJob
+        :type other: ClassificationJob
+        :return: True or False
+        :rtype: bool
+        """
         if not isinstance(other, ClassificationJob):
             return NotImplemented
 
-        if not super(ClassificationJob, self).__eq__(other):
+        if not super().__eq__(other):
             return False
 
         return self.primary_metric == other.primary_metric
 
     def __ne__(self, other):
+        """Check inequality between two ImageLimitSettings objects.
+
+        :param other: An instance of ClassificationJob
+        :type other: ClassificationJob
+        :return: True or False
+        :rtype: bool
+        """
         return not self.__eq__(other)
