@@ -33,10 +33,17 @@ if __name__ == "__main__":
         required=True,
     )
 
+    parser.add_argument(
+        "--next",
+        default=False,
+        help="Next version of pyright is being tested.",
+        required=False
+    )
+
     args = parser.parse_args()
     package_name = os.path.basename(os.path.abspath(args.target_package))
 
-    if in_ci():
+    if not args.next and in_ci():
         if not is_check_enabled(args.target_package, "pyright") or is_typing_ignored(package_name):
             logging.info(
                 f"Package {package_name} opts-out of pyright check. See https://aka.ms/python/typing-guide for information."
@@ -48,7 +55,7 @@ if __name__ == "__main__":
         os.path.join(args.target_package, "samples"),
     ]
 
-    if in_ci():
+    if not args.next and in_ci():
         if not is_check_enabled(args.target_package, "type_check_samples"):
             logging.info(
                 f"Package {package_name} opts-out of pyright check on samples."
