@@ -8,11 +8,11 @@ import warnings
 
 from opentelemetry import trace
 from opentelemetry.trace import Span, Tracer, SpanKind as OpenTelemetrySpanKind, Link as OpenTelemetryLink
-from opentelemetry.context import attach, detach, get_current
-from opentelemetry.propagate import extract, inject
-from opentelemetry.trace.propagation import get_current_span as get_span_from_context
+from opentelemetry.context import attach, detach, get_current  # type: ignore[attr-defined]
+from opentelemetry.propagate import extract, inject  # type: ignore[attr-defined]
+from opentelemetry.trace.propagation import get_current_span as get_span_from_context  # type: ignore[attr-defined]
 
-from azure.core.tracing import SpanKind, HttpSpanMixin  # pylint: disable=no-name-in-module
+from azure.core.tracing import SpanKind, HttpSpanMixin  # type: ignore[attr-defined] # pylint: disable=no-name-in-module
 
 from ._schema import OpenTelemetrySchema
 from ._version import VERSION
@@ -40,13 +40,13 @@ class OpenTelemetrySpan(HttpSpanMixin, object):
     :type span: ~OpenTelemetry.trace.Span
     :param name: The name of the OpenTelemetry span to create if a new span is needed
     :type name: str
-    :keyword SpanKind kind: The span kind of this span.
+    :keyword kind: The span kind of this span.
+    :paramtype kind: ~azure.core.tracing.SpanKind
     :keyword links: The list of links to be added to the span.
     :paramtype links: list[~azure.core.tracing.Link]
     """
 
-    def __init__(self, span=None, name="span", **kwargs):
-        # type: (Optional[Span], Optional[str], Any) -> None
+    def __init__(self, span: Optional[Span] = None, name: str = "span", **kwargs: Any) -> None:
         current_tracer = self.get_current_tracer()
 
         # TODO: Once we have additional supported versions, we should add a way to specify the version.
@@ -69,7 +69,7 @@ class OpenTelemetrySpan(HttpSpanMixin, object):
             else OpenTelemetrySpanKind.INTERNAL
             if value == SpanKind.UNSPECIFIED
             else None
-        )  # type: SpanKind
+        )
         if value and kind is None:
             raise ValueError("Kind {} is not supported in OpenTelemetry".format(value))
 
@@ -83,27 +83,26 @@ class OpenTelemetrySpan(HttpSpanMixin, object):
                     ot_links.append(OpenTelemetryLink(span_ctx, link.attributes))
                 kwargs.setdefault("links", ot_links)
             except AttributeError:
-                # we will just send the links as is if it's not ~azure.core.tracing.Link without any validation
+                # We will just send the links as is if it's not ~azure.core.tracing.Link without any validation
                 # assuming user knows what they are doing.
                 kwargs.setdefault("links", links)
-        self._span_instance = span or current_tracer.start_span(name=name, kind=kind, **kwargs)
-        self._current_ctxt_manager = None
+        self._span_instance = span or current_tracer.start_span(name=name, kind=kind, **kwargs)  # type: ignore
+        self._current_ctxt_manager: Optional[ContextManager[Span]] = None
 
     @property
-    def span_instance(self):
-        # type: () -> Span
+    def span_instance(self) -> Span:
         """
         :return: The OpenTelemetry span that is being wrapped.
         """
         return self._span_instance
 
-    def span(self, name="span", **kwargs):
-        # type: (Optional[str], Any) -> OpenTelemetrySpan
+    def span(self, name: str = "span", **kwargs: Any) -> "OpenTelemetrySpan":
         """
         Create a child span for the current span and append it to the child spans list in the span instance.
         :param name: Name of the child span
         :type name: str
-        :keyword SpanKind kind: The span kind of this span.
+        :keyword kind: The span kind of this span.
+        :paramtype kind: ~azure.core.tracing.SpanKind
         :keyword links: The list of links to be added to the span.
         :paramtype links: list[Link]
         :return: The OpenTelemetrySpan that is wrapping the child span instance
@@ -111,10 +110,9 @@ class OpenTelemetrySpan(HttpSpanMixin, object):
         return self.__class__(name=name, **kwargs)
 
     @property
-    def kind(self):
-        # type: () -> Optional[SpanKind]
+    def kind(self) -> Optional[SpanKind]:
         """Get the span kind of this span."""
-        value = self.span_instance.kind
+        value = self.span_instance.kind  # type: ignore[attr-defined]
         return (
             SpanKind.CLIENT
             if value == OpenTelemetrySpanKind.CLIENT
@@ -130,8 +128,7 @@ class OpenTelemetrySpan(HttpSpanMixin, object):
         )
 
     @kind.setter
-    def kind(self, value):
-        # type: (SpanKind) -> None
+    def kind(self, value: SpanKind) -> None:
         """Set the span kind of this span."""
         kind = (
             OpenTelemetrySpanKind.CLIENT
@@ -151,7 +148,7 @@ class OpenTelemetrySpan(HttpSpanMixin, object):
         if kind is None:
             raise ValueError("Kind {} is not supported in OpenTelemetry".format(value))
         try:
-            self._span_instance._kind = kind  # pylint: disable=protected-access
+            self._span_instance._kind = kind  # type: ignore[attr-defined] # pylint: disable=protected-access
         except AttributeError:
             warnings.warn(
                 """Kind must be set while creating the span for OpenTelemetry. It might be possible
@@ -159,52 +156,50 @@ class OpenTelemetrySpan(HttpSpanMixin, object):
                 Try updating the azure packages to the latest versions."""
             )
 
-    def __enter__(self):
+    def __enter__(self) -> "OpenTelemetrySpan":
         """Start a span."""
         self._current_ctxt_manager = trace.use_span(self._span_instance, end_on_exit=True)
-        self._current_ctxt_manager.__enter__()  # pylint: disable=no-member
+        if self._current_ctxt_manager:
+            self._current_ctxt_manager.__enter__()  # pylint: disable=no-member
         return self
 
-    def __exit__(self, exception_type, exception_value, traceback):
+    def __exit__(self, exception_type, exception_value, traceback) -> None:
         """Finish a span."""
         if self._current_ctxt_manager:
             self._current_ctxt_manager.__exit__(exception_type, exception_value, traceback)  # pylint: disable=no-member
             self._current_ctxt_manager = None
 
-    def start(self):
-        # Spans are automatically started at their creation with OpenTelemetry
+    def start(self) -> None:
+        # Spans are automatically started at their creation with OpenTelemetry.
         pass
 
-    def finish(self):
-        # type: () -> None
+    def finish(self) -> None:
         """Set the end time for a span."""
         self.span_instance.end()
 
-    def to_header(self):  # pylint: disable=no-self-use
-        # type: () -> Dict[str, str]
+    def to_header(self) -> Dict[str, str]:  # pylint: disable=no-self-use
         """
         Returns a dictionary with the header labels and values.
         :return: A key value pair dictionary
+        :rtype: dict[str, str]
         """
-        temp_headers = {}  # type: Dict[str, str]
+        temp_headers: Dict[str, str] = {}
         inject(temp_headers)
         return temp_headers
 
-    def add_attribute(self, key, value):
-        # type: (str, Union[str, int]) -> None
+    def add_attribute(self, key: str, value: Union[str, int]) -> None:
         """
         Add attribute (key value pair) to the current span.
 
         :param key: The key of the key value pair
         :type key: str
         :param value: The value of the key value pair
-        :type value: str
+        :type value: Union[str, int]
         """
         key = self._attribute_mappings.get(key, key)
         self.span_instance.set_attribute(key, value)
 
-    def get_trace_parent(self):
-        # type: () -> str
+    def get_trace_parent(self) -> str:
         """Return traceparent string as defined in W3C trace context specification.
 
         Example:
@@ -220,8 +215,7 @@ class OpenTelemetrySpan(HttpSpanMixin, object):
         return self.to_header()["traceparent"]
 
     @classmethod
-    def link(cls, traceparent, attributes=None):
-        # type: (str, Attributes) -> None
+    def link(cls, traceparent: str, attributes: Attributes = None) -> None:
         """
         Links the context to the current tracer.
 
@@ -231,8 +225,7 @@ class OpenTelemetrySpan(HttpSpanMixin, object):
         cls.link_from_headers({"traceparent": traceparent}, attributes)
 
     @classmethod
-    def link_from_headers(cls, headers, attributes=None):
-        # type: (Dict[str, str], Attributes) -> None
+    def link_from_headers(cls, headers: Dict[str, str], attributes: Attributes = None):
         """
         Given a dictionary, extracts the context and links the context to the current tracer.
 
@@ -243,7 +236,7 @@ class OpenTelemetrySpan(HttpSpanMixin, object):
         span_ctx = get_span_from_context(ctx).get_span_context()
         current_span = cls.get_current_span()
         try:
-            current_span._links.append(OpenTelemetryLink(span_ctx, attributes))  # pylint: disable=protected-access
+            current_span._links.append(OpenTelemetryLink(span_ctx, attributes))  # type: ignore # pylint: disable=protected-access
         except AttributeError:
             warnings.warn(
                 """Link must be added while creating the span for OpenTelemetry. It might be possible
@@ -252,48 +245,42 @@ class OpenTelemetrySpan(HttpSpanMixin, object):
             )
 
     @classmethod
-    def get_current_span(cls):
-        # type: () -> Span
+    def get_current_span(cls) -> Span:
         """
-        Get the current span from the execution context. Return None otherwise.
+        Get the current span from the execution context.
         """
         return get_span_from_context()
 
     @classmethod
-    def get_current_tracer(cls):
-        # type: () -> Tracer
+    def get_current_tracer(cls) -> Tracer:
         """
-        Get the current tracer from the execution context. Return None otherwise.
+        Get the current tracer from the execution context.
         """
         return trace.get_tracer(__name__, __version__)
 
     @classmethod
-    def change_context(cls, span):
-        # type: (Span) -> ContextManager
+    def change_context(cls, span: Span) -> ContextManager:
         """Change the context for the life of this context manager."""
         return trace.use_span(span, end_on_exit=False)
 
     @classmethod
-    def set_current_span(cls, span):
-        # type: (Span) -> None
+    def set_current_span(cls, span: Span) -> None:
         """Not supported by OpenTelemetry."""
         raise NotImplementedError(
             "set_current_span is not supported by OpenTelemetry plugin. Use change_context instead."
         )
 
     @classmethod
-    def set_current_tracer(cls, _):
-        # type: (Tracer) -> None
-        """
-        Set the given tracer as the current tracer in the execution context.
+    def set_current_tracer(cls, _: Tracer) -> None:
+        """Set the given tracer as the current tracer in the execution context.
+
         :param tracer: The tracer to set the current tracer as
         :type tracer: :class: OpenTelemetry.trace.Tracer
         """
         # Do nothing, if you're able to get two tracer with OpenTelemetry that's a surprise!
 
     @classmethod
-    def with_current_context(cls, func):
-        # type: (Callable) -> Callable
+    def with_current_context(cls, func: Callable) -> Callable:
         """Passes the current spans to the new context the function will be run in.
 
         :param func: The function that will be run in the new context
