@@ -4,8 +4,8 @@ with the Eventhub SDK.
 
 This example traces calls for sending a batch to eventhub.
 
-An alternative path to export using AzureMonitor is also mentioned in the sample. Please take
-a look at the commented code.
+An alternative path to export using the OpenTelemetry exporter for Azure Monitor
+is also mentioned in the sample. Please take a look at the commented code.
 """
 
 # Declare OpenTelemetry as enabled tracing plugin for Azure SDKs
@@ -15,12 +15,14 @@ from azure.core.tracing.ext.opentelemetry_span import OpenTelemetrySpan
 settings.tracing_implementation = OpenTelemetrySpan
 
 # In the below example, we use a simple console exporter, uncomment these lines to use
-# the Azure Monitor Exporter. It can be installed from https://pypi.org/project/opentelemetry-azure-monitor/
-# Example of Azure Monitor exporter, but you can use anything OpenTelemetry supports
-# from azure_monitor import AzureMonitorSpanExporter
-# exporter = AzureMonitorSpanExporter(
-#     instrumentation_key="uuid of the instrumentation key (see your Azure Monitor account)"
+# the OpenTelemetry exporter for Azure Monitor.
+# Example of a trace exporter for Azure Monitor, but you can use anything OpenTelemetry supports.
+
+# from azure.monitor.opentelemetry.exporter import AzureMonitorTraceExporter
+# exporter = AzureMonitorTraceExporter(
+#     connection_string="the connection string used for your Application Insights resource"
 # )
+
 
 # Regular open telemetry usage from here, see https://github.com/open-telemetry/opentelemetry-python
 # for details
