@@ -27,7 +27,7 @@
 
 import functools
 
-from typing import Callable, Any, TypeVar, overload
+from typing import Callable, Any, TypeVar, overload, Optional
 from typing_extensions import ParamSpec
 from .common import change_context, get_function_and_class_name
 from . import SpanKind as _SpanKind
@@ -50,7 +50,7 @@ def distributed_trace(  # pylint:disable=function-redefined
     pass
 
 
-def distributed_trace(__func: Callable[P, T] = None, **kwargs: Any):  # pylint:disable=function-redefined
+def distributed_trace(__func: Optional[Callable[P, T]] = None, **kwargs: Any):  # pylint:disable=function-redefined
     """Decorator to apply to function to get traced automatically.
 
     Span will use the func name or "name_of_span".
