@@ -28,7 +28,7 @@ def generate_enum_content(tuples):
         print(k + " = '" + v + "'\n")
     print("# Aliases end here")
     for tup in tup_list:
-        print(tup[0] + " = '" + tup[1].replace('API', 'Api') + "'\n")
+        print(tup[0] + " = '" + tup[1] + "'\n")
     for k, v in additional_events.items():
         print(k + " = '" + v + "'\n")
 
