@@ -8,7 +8,7 @@ from typing import Dict
 
 from marshmallow import fields, post_load
 
-from azure.ai.ml._restclient.v2022_01_01_preview.models import ConnectionAuthType
+from azure.ai.ml._restclient.v2022_12_01_preview.models import ConnectionAuthType
 from azure.ai.ml._schema.core.fields import StringTransformedEnum
 from azure.ai.ml._schema.core.schema import PatchedSchemaMeta
 from azure.ai.ml._utils.utils import camel_to_snake
@@ -18,6 +18,7 @@ from azure.ai.ml.entities._credentials import (
     SasTokenConfiguration,
     ServicePrincipalConfiguration,
     UsernamePasswordConfiguration,
+    AccessKeyConfiguration,
 )
 
 
@@ -98,3 +99,18 @@ class ServicePrincipalConfigurationSchema(metaclass=PatchedSchemaMeta):
     def make(self, data: Dict[str, str], **kwargs) -> ServicePrincipalConfiguration:
         data.pop("type")
         return ServicePrincipalConfiguration(**data)
+
+
+class AccessKeyCredentialsSchema(metaclass=PatchedSchemaMeta):
+    type = StringTransformedEnum(
+        allowed_values=ConnectionAuthType.ACCESS_KEY,
+        casing_transform=camel_to_snake,
+        required=True,
+    )
+    access_key_id = fields.Str()
+    secret_access_key = fields.Str()
+
+    @post_load
+    def make(self, data: Dict[str, str], **kwargs) -> AccessKeyConfiguration:
+        data.pop("type")
+        return AccessKeyConfiguration(**data)
