@@ -2,7 +2,8 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 
-# pylint: disable=too-many-instance-attributes
+# pylint: disable=too-many-instance-attributes,protected-access
+
 
 from os import PathLike
 from pathlib import Path
@@ -12,7 +13,9 @@ from marshmallow import ValidationError
 from azure.ai.ml._restclient.v2022_12_01_preview.models import Workspace as RestWorkspace
 
 from azure.ai.ml._schema._feature_store.feature_store_schema import FeatureStoreSchema
-from azure.ai.ml.entities import Workspace, CustomerManagedKey, FeatureStoreSettings, ComputeRuntime
+from azure.ai.ml.entities._workspace.feature_store_settings import FeatureStoreSettings
+from azure.ai.ml.entities._workspace.compute_runtime import ComputeRuntime
+from azure.ai.ml.entities import Workspace, CustomerManagedKey
 from azure.ai.ml.entities._util import load_from_dict
 from azure.ai.ml.entities._credentials import IdentityConfiguration, ManagedIdentityConfiguration
 from azure.ai.ml._utils._experimental import experimental
@@ -144,15 +147,15 @@ class FeatureStore(Workspace):
         if not rest_obj:
             return None
 
-        workspace_object = Workspace._from_rest_object(rest_obj)  # pylint: disable=protected-access
+        workspace_object = Workspace._from_rest_object(rest_obj)
 
         return FeatureStore(
             name=workspace_object.name,
             description=workspace_object.description,
             tags=workspace_object.tags,
-            compute_runtime=ComputeRuntime._from_rest_object(  # pylint: disable=protected-access
-                workspace_object.feature_store_settings.compute_runtime
-                if workspace_object.feature_store_settings
+            compute_runtime=ComputeRuntime._from_rest_object(
+                workspace_object._feature_store_settings.compute_runtime
+                if workspace_object._feature_store_settings
                 else None
             ),
             display_name=workspace_object.display_name,
