@@ -6,9 +6,10 @@ __path__ = __import__("pkgutil").extend_path(__path__, __name__)
 
 from .search_space import SearchSpace
 from .stack_ensemble_settings import StackEnsembleSettings
-from .training_settings import TrainingSettings
+from .training_settings import ClassificationTrainingSettings, TrainingSettings
 
 __all__ = [
+    "ClassificationTrainingSettings",
     "TrainingSettings",
     "SearchSpace",
     "StackEnsembleSettings",
