@@ -10,14 +10,66 @@ from azure.ai.ml.entities._mixins import RestTranslatableMixin
 
 
 class ImageLimitSettings(RestTranslatableMixin):
-    """Limit settings for all AutoML Image Verticals.
+    r"""Limit settings for AutoML Image Verticals.
 
-    :param max_concurrent_trials: Maximum number of concurrent AutoML iterations.
-    :type max_concurrent_trials: int, optional
-    :param max_trials: Maximum number of AutoML iterations.
-    :type max_trials: int, optional
-    :param timeout_minutes: AutoML job timeout.
-    :type timeout_minutes: int, optional
+    ImageLimitSettings is a class that contains the following parameters:  max_concurrent_trials, max_trials, and \
+    timeout_minutes. It is used to configure the limits for the AutoML job.
+
+    .. remarks::
+
+        ImageLimitSettings is an optional configuration method to configure limits parameters such as timeouts etc.
+
+            .. note::
+
+                The number of concurrent runs is gated on the resources available in the specified compute target.
+                Ensure that the compute target has the available resources for the desired concurrency.
+
+    :keyword max_concurrent_trials: Maximum number of concurrent AutoML iterations, defaults to None.
+
+        .. note::
+
+            Do not set the value of max_trials as a negative integer.
+
+    :paramtype  max_concurrent_trials: typing.Optional[int]
+    :keyword max_trials: Represents the maximum number of trials (children jobs) that would be executed in parallel.
+    :paramtype  max_trials: typing.Optional[int]
+    :keyword timeout_minutes: AutoML job timeout, defaults to None
+    :paramtype  timeout_minutes: typing.Optional[int]
+
+    :raises ValueError: If max_concurrent_trials is not None and is not a positive integer.
+    :raises ValueError: If max_trials is not None and is not a positive integer.
+    :raises ValueError: If timeout_minutes is not None and is not a positive integer.
+    :return: ImageLimitSettings object.
+    :rtype: ImageLimitSettings
+
+    .. tip::
+        It's a good practice to match max_concurrent_trials count with the number of nodes in the cluster.
+        For example, if you have a cluster with 4 nodes, set max_concurrent_trials to 4.
+
+    **Example usage**
+
+    .. code-block:: python
+        :caption: Configuration of ImageLimitSettings
+
+        from azure.ai.ml import automl
+
+        # Create the AutoML job with the related factory-function.
+        image_instance_segmentation_job = automl.image_instance_segmentation(
+            compute=compute_name,
+            experiment_name=exp_name,
+            training_data=my_training_data_input,
+            validation_data=my_validation_data_input,
+            target_column_name="label",
+            primary_metric="MeanAveragePrecision",
+            tags={"my_custom_tag": "custom value"},
+        )
+        # Set the limits for the AutoML job.
+        image_instance_segmentation_job.set_limits(
+            max_trials=10,
+            max_concurrent_trials=2,
+        )
+        # Submit the AutoML job.
+        image_instance_segmentation_job.submit()
     """
 
     def __init__(
@@ -27,11 +79,34 @@ class ImageLimitSettings(RestTranslatableMixin):
         max_trials: Optional[int] = None,
         timeout_minutes: Optional[int] = None,
     ):
+        """Initialize an ImageLimitSettings object.
+
+        Constructor for ImageLimitSettings for AutoML Image Verticals.
+
+        :keyword  max_concurrent_trials: Represents the maximum number of trials (children jobs) that would be \
+            executed in parallel.
+        :paramtype max_concurrent_trials: typing.Optional[int]
+        :keyword max_trials: Maximum number of AutoML iterations, defaults to None
+        :paramtype max_trials: typing.Optional[int]
+        :keyword timeout_minutes: AutoML job timeout, defaults to None
+        :paramtype timeout_minutes: typing.Optional[int]
+        :raises ValueError: If max_concurrent_trials is not None and is not a positive integer.
+        :raises ValueError: If max_trials is not None and is not a positive integer.
+        :raises ValueError: If timeout_minutes is not None and is not a positive integer.
+
+        :return: ImageLimitSettings object.
+        :rtype: ImageLimitSettings
+        """
         self.max_concurrent_trials = max_concurrent_trials
         self.max_trials = max_trials
         self.timeout_minutes = timeout_minutes
 
     def _to_rest_object(self) -> RestImageLimitSettings:
+        """Convert ImageLimitSettings objects to a rest object.
+
+        :return: A rest object of ImageLimitSettings objects.
+        :rtype: RestImageLimitSettings
+        """
         return RestImageLimitSettings(
             max_concurrent_trials=self.max_concurrent_trials,
             max_trials=self.max_trials,
@@ -40,6 +115,13 @@ class ImageLimitSettings(RestTranslatableMixin):
 
     @classmethod
     def _from_rest_object(cls, obj: RestImageLimitSettings) -> "ImageLimitSettings":
+        """Convert the rest object to a dict containing items to init the ImageLimitSettings objects.
+
+        :param obj: Limit settings for the AutoML job in Rest format.
+        :type obj: RestImageLimitSettings
+        :return: Limit settings for an AutoML Image Vertical.
+        :rtype: ImageLimitSettings
+        """
         return cls(
             max_concurrent_trials=obj.max_concurrent_trials,
             max_trials=obj.max_trials,
@@ -47,6 +129,15 @@ class ImageLimitSettings(RestTranslatableMixin):
         )
 
     def __eq__(self, other: object) -> bool:
+        """Check equality between two ImageLimitSettings objects.
+
+        _extended_summary_
+
+        :param other: ImageLimitSettings object
+        :type other: object
+        :return: True or False
+        :rtype: bool
+        """
         if not isinstance(other, ImageLimitSettings):
             return NotImplemented
 
@@ -57,4 +148,11 @@ class ImageLimitSettings(RestTranslatableMixin):
         )
 
     def __ne__(self, other: object) -> bool:
+        """Check inequality between two ImageLimitSettings objects.
+
+        :param other: ImageLimitSettings object
+        :type other: object
+        :return: True or False
+        :rtype: bool
+        """
         return not self.__eq__(other)
