@@ -10,7 +10,6 @@ from pathlib import Path
 from typing import IO, Any, AnyStr, Dict, Optional, Union
 
 from azure.ai.ml._restclient.v2022_01_01_preview.models import (
-    ConnectionAuthType,
     ManagedIdentityAuthTypeWorkspaceConnectionProperties,
     NoneAuthTypeWorkspaceConnectionProperties,
     PATAuthTypeWorkspaceConnectionProperties,
@@ -20,6 +19,10 @@ from azure.ai.ml._restclient.v2022_01_01_preview.models import (
 )
 from azure.ai.ml._restclient.v2022_01_01_preview.models import (
     WorkspaceConnectionPropertiesV2BasicResource as RestWorkspaceConnection,
+)
+from azure.ai.ml._restclient.v2022_12_01_preview.models import (
+    ConnectionAuthType,
+    AccessKeyAuthTypeWorkspaceConnectionProperties,
 )
 from azure.ai.ml._schema.workspace.connections.workspace_connection import WorkspaceConnectionSchema
 from azure.ai.ml._utils._experimental import experimental
@@ -31,6 +34,7 @@ from azure.ai.ml.entities._credentials import (
     SasTokenConfiguration,
     ServicePrincipalConfiguration,
     UsernamePasswordConfiguration,
+    AccessKeyConfiguration,
 )
 from azure.ai.ml.entities._resource import Resource
 from azure.ai.ml.entities._system_data import SystemData
@@ -50,11 +54,13 @@ class WorkspaceConnection(Resource):
     :type credentials: Union[
         ~azure.ai.ml.entities.PatTokenConfiguration, ~azure.ai.ml.entities.SasTokenConfiguration,
         ~azure.ai.ml.entities.UsernamePasswordConfiguration, ~azure.ai.ml.entities.ManagedIdentityConfiguration
-        ~azure.ai.ml.entities.ServicePrincipalConfiguration
+        ~azure.ai.ml.entities.ServicePrincipalConfiguration, ~azure.ai.ml.entities.AccessKeyConfiguration
         ]
     :param type: The category of external resource for this connection.
-    :type type: The type of workspace connection, possible values are:
-        ["git", "python_feed", "container_registry", "feature_store"]
+    :type type: The type of workspace connection, possible values are: [
+        "git", "python_feed", "container_registry", "feature_store", "s3", "snowflake",
+         "azure_sql_db", "azure_synapse_analytics", "azure_my_sql_db", "azure_postgres_db"
+          ]
     """
 
     def __init__(
@@ -69,6 +75,7 @@ class WorkspaceConnection(Resource):
             UsernamePasswordConfiguration,
             ManagedIdentityConfiguration,
             ServicePrincipalConfiguration,
+            AccessKeyConfiguration,
         ],
         metadata: Optional[Dict[str, Any]] = None,
         **kwargs,
@@ -112,6 +119,7 @@ class WorkspaceConnection(Resource):
         UsernamePasswordConfiguration,
         ManagedIdentityConfiguration,
         ServicePrincipalConfiguration,
+        AccessKeyConfiguration,
     ]:
         """Credentials for workspace connection.
 
@@ -122,6 +130,7 @@ class WorkspaceConnection(Resource):
             UsernamePasswordCredentialsConfiguration,
             ManagedIdentityConfiguration,
             ServicePrincipalCredentialsConfiguration,
+            AccessKeyCredentialsConfiguration,
         ]
         """
         return self._credentials
@@ -190,7 +199,8 @@ class WorkspaceConnection(Resource):
             credentials = ManagedIdentityConfiguration._from_workspace_connection_rest_object(properties.credentials)
         if properties.auth_type == ConnectionAuthType.USERNAME_PASSWORD:
             credentials = UsernamePasswordConfiguration._from_workspace_connection_rest_object(properties.credentials)
-
+        if properties.auth_type == ConnectionAuthType.ACCESS_KEY:
+            credentials = AccessKeyConfiguration._from_workspace_connection_rest_object(properties.credentials)
         if properties.auth_type == ConnectionAuthType.SERVICE_PRINCIPAL:
             credentials = ServicePrincipalConfiguration._from_workspace_connection_rest_object(properties.credentials)
 
@@ -201,7 +211,7 @@ class WorkspaceConnection(Resource):
             creation_context=SystemData._from_rest_object(rest_obj.system_data) if rest_obj.system_data else None,
             type=camel_to_snake(properties.category),
             credentials=credentials,
-            metadata=properties.metadata,
+            metadata=properties.metadata if hasattr(properties, "metadata") else None,
         )
 
         return workspace_connection
@@ -219,6 +229,8 @@ class WorkspaceConnection(Resource):
             workspace_connection_properties_class = ManagedIdentityAuthTypeWorkspaceConnectionProperties
         elif auth_type == camel_to_snake(ConnectionAuthType.USERNAME_PASSWORD):
             workspace_connection_properties_class = UsernamePasswordAuthTypeWorkspaceConnectionProperties
+        elif auth_type == camel_to_snake(ConnectionAuthType.ACCESS_KEY):
+            workspace_connection_properties_class = AccessKeyAuthTypeWorkspaceConnectionProperties
         elif auth_type == camel_to_snake(ConnectionAuthType.SAS):
             workspace_connection_properties_class = SASAuthTypeWorkspaceConnectionProperties
         elif auth_type == camel_to_snake(ConnectionAuthType.SERVICE_PRINCIPAL):
