@@ -9,7 +9,6 @@ from pathlib import Path
 from typing import IO, AnyStr, Dict, Optional, Union
 
 from azure.ai.ml._restclient.v2022_12_01_preview.models import ManagedServiceIdentity as RestManagedServiceIdentity
-from azure.ai.ml._restclient.v2022_12_01_preview.models import FeatureStoreSettings as RestFeatureStoreSettings
 from azure.ai.ml._restclient.v2022_12_01_preview.models import Workspace as RestWorkspace
 from azure.ai.ml._restclient.v2022_12_01_preview.models import ManagedNetworkSettings as RestManagedNetwork
 from azure.ai.ml._schema.workspace.workspace import WorkspaceSchema
@@ -46,7 +45,6 @@ class Workspace(Resource):
         identity: Optional[IdentityConfiguration] = None,
         primary_user_assigned_identity: Optional[str] = None,
         managed_network: Optional[ManagedNetwork] = None,
-        feature_store_settings: Optional[FeatureStoreSettings] = None,
         **kwargs,
     ):
 
@@ -97,8 +95,6 @@ class Workspace(Resource):
         :type primary_user_assigned_identity: str
         :param managed_network: workspace's Managed Network configuration
         :type managed_network: ManagedNetwork
-        :param feature_store_settings: Settings for feature store type workspace.
-        :type feature_store_settings: FeatureStoreSettings
         :param kwargs: A dictionary of additional configuration parameters.
         :type kwargs: dict
         """
@@ -120,7 +116,7 @@ class Workspace(Resource):
         self.public_network_access = public_network_access
         self.identity = identity
         self.primary_user_assigned_identity = primary_user_assigned_identity
-        self.feature_store_settings = feature_store_settings
+        self._feature_store_settings: Optional[FeatureStoreSettings] = kwargs.get("feature_store_settings", None)
         self.managed_network = managed_network
 
     @property
@@ -211,11 +207,6 @@ class Workspace(Resource):
             identity = IdentityConfiguration._from_workspace_rest_object(  # pylint: disable=protected-access
                 rest_obj.identity
             )
-        feature_store_settings = None
-        if rest_obj.feature_store_settings and isinstance(rest_obj.feature_store_settings, RestFeatureStoreSettings):
-            feature_store_settings = FeatureStoreSettings._from_rest_object(  # pylint: disable=protected-access
-                rest_obj.feature_store_settings
-            )
         return Workspace(
             name=rest_obj.name,
             id=rest_obj.id,
@@ -238,13 +229,9 @@ class Workspace(Resource):
             identity=identity,
             primary_user_assigned_identity=rest_obj.primary_user_assigned_identity,
             managed_network=managed_network,
-            feature_store_settings=feature_store_settings,
         )
 
     def _to_rest_object(self) -> RestWorkspace:
-        feature_store_Settings = None
-        if self.feature_store_settings:
-            feature_store_Settings = self.feature_store_settings._to_rest_object()  # pylint: disable=protected-access
 
         return RestWorkspace(
             identity=self.identity._to_workspace_rest_object()  # pylint: disable=protected-access
@@ -264,7 +251,6 @@ class Workspace(Resource):
             image_build_compute=self.image_build_compute,
             public_network_access=self.public_network_access,
             primary_user_assigned_identity=self.primary_user_assigned_identity,
-            feature_store_settings=feature_store_Settings,
             managed_network=self.managed_network._to_rest_object()  # pylint: disable=protected-access
             if self.managed_network
             else None,  # pylint: disable=protected-access
