@@ -16,13 +16,13 @@ from azure.core.credentials import TokenCredential
 from azure.core.polling import LROPoller
 from azure.core.tracing.decorator import distributed_trace
 from azure.ai.ml._utils._logger_utils import OpsLogger
+from azure.ai.ml.entities._feature_store.feature_store import FeatureStore
+from azure.ai.ml.entities._workspace.feature_store_settings import FeatureStoreSettings
+from azure.ai.ml.entities._feature_store.materialization_store import MaterializationStore
 from azure.ai.ml.entities import (
-    FeatureStore,
-    FeatureStoreSettings,
     ManagedIdentityConfiguration,
     IdentityConfiguration,
     WorkspaceConnection,
-    MaterializationStore,
 )
 from azure.ai.ml._utils._experimental import experimental
 from azure.ai.ml.constants._common import Scope
