@@ -755,7 +755,7 @@ class TestCommandFunction:
         )
         result = node._validate()
         message = "Should not specify min or max executors when dynamic allocation is disabled."
-        assert "conf" in result.error_messages and message == result.error_messages["conf"]
+        assert "*" in result.error_messages and message == result.error_messages["*"]
 
     def test_executor_instances_is_mandatory_when_dynamic_allocation_disabled(self):
         node = spark(
@@ -772,7 +772,7 @@ class TestCommandFunction:
             "spark.driver.cores, spark.driver.memory, spark.executor.cores, spark.executor.memory and "
             "spark.executor.instances are mandatory fields."
         )
-        assert "conf" in result.error_messages and message == result.error_messages["conf"]
+        assert "*" in result.error_messages and message == result.error_messages["*"]
 
     def test_executor_instances_is_specified_as_min_executor_if_unset(self):
         node = spark(
@@ -809,7 +809,7 @@ class TestCommandFunction:
             "Executor instances must be a valid non-negative integer and must be between "
             "spark.dynamicAllocation.minExecutors and spark.dynamicAllocation.maxExecutors"
         )
-        assert "conf" in result.error_messages and message == result.error_messages["conf"]
+        assert "*" in result.error_messages and message == result.error_messages["*"]
 
     def test_spark_job_with_additional_conf(self):
         node = spark(
